requirements.txt
aim==3.27.0
aim-ui==3.27.0
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==25.1.0
base58==2.0.1
boto3==1.36.17
botocore==1.36.17
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
docker-pycreds==0.4.0
fastapi==0.115.8
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
llvmlite==0.44.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
RestrictedPython==8.0
rich==13.9.4
s3transfer==0.11.2
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
torch==2.6.0+cpu
torch-geometric==2.6.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.6
websockets==14.2
yarl==1.18.3